Added event creation *
* needs to use datetime object for start/end times

moved into a backend folder to match jakes branch and avoid potential conflicts (i hope)
